requirements.txt
Pillow==8.2.0
psycopg2==2.8.6
sqlparse==0.4.1
djangorestframework==3.12.4
xmltodict==0.12.0
urllib3==1.26.4

# Django
# -----------------------------------------
Django~=3.1.5
django-environ==0.4.5
django-model-utils==4.1.1
django-redis==4.12.1
django-filter==2.4.0
dj-database-url==0.5.0
redis==3.5.3
celery==5.0.5
django-extensions==3.1.3
ipython==7.23.1
drf-yasg==1.20.0
django-rest-swagger==2.2.0
